Moved repository classes to a new package under repository package
