+++ mosad_backend/src/main/java/org/rtss/mosad_backend/service/stock_management_service/ItemBatteryService.java
@@ -4,7 +4,7 @@ import jakarta.transaction.Transactional;
 import org.rtss.mosad_backend.dto.stock_management_dto.ItemBatteryDTO;
 import org.rtss.mosad_backend.entity.stock_management_entity.ItemBattery;
 import org.rtss.mosad_backend.model_mapper.stock_management_model_mapper.ItemBatteryDTOMapper;
-import org.rtss.mosad_backend.repository.ItemBatteryRepo;
+import org.rtss.mosad_backend.repository.stock_management_repository.ItemBatteryRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
