+++ mosad_backend/src/main/java/org/rtss/mosad_backend/service/stock_management_service/ItemOtherAccessoriesService.java
@@ -2,10 +2,9 @@ package org.rtss.mosad_backend.service.stock_management_service;
 
 import jakarta.transaction.Transactional;
 import org.rtss.mosad_backend.dto.stock_management_dto.ItemOtherAccessoriesDTO;
-import org.rtss.mosad_backend.dto.stock_management_dto.ItemOtherAccessoriesDTO;
 import org.rtss.mosad_backend.entity.stock_management_entity.ItemOtherAccessories;
 import org.rtss.mosad_backend.model_mapper.stock_management_model_mapper.ItemOtherAccessoriesDTOMapper;
-import org.rtss.mosad_backend.repository.ItemOtherAccessoriesRepo;
+import org.rtss.mosad_backend.repository.stock_management_repository.ItemOtherAccessoriesRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
