+++ mosad_backend/src/main/java/org/rtss/mosad_backend/service/stock_management_service/ItemTyreLinglongService.java
@@ -11,10 +11,9 @@ package org.rtss.mosad_backend.service.stock_management_service;
 
 import jakarta.transaction.Transactional;
 import org.rtss.mosad_backend.dto.stock_management_dto.ItemTyreLinglongDTO;
-import org.rtss.mosad_backend.dto.stock_management_dto.ItemTyreLinglongDTO;
 import org.rtss.mosad_backend.entity.stock_management_entity.ItemTyreLinglong;
 import org.rtss.mosad_backend.model_mapper.stock_management_model_mapper.ItemTyreLinglongDTOMapper;
-import org.rtss.mosad_backend.repository.ItemTyreLinglongRepo;
+import org.rtss.mosad_backend.repository.stock_management_repository.ItemTyreLinglongRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
