+++ mosad_backend/src/main/java/org/rtss/mosad_backend/service/stock_management_service/ItemTyrePresaService.java
@@ -1,10 +1,9 @@
 package org.rtss.mosad_backend.service.stock_management_service;
 
 import org.rtss.mosad_backend.dto.stock_management_dto.ItemTyrePresaDTO;
-import org.rtss.mosad_backend.dto.stock_management_dto.ItemTyrePresaDTO;
 import org.rtss.mosad_backend.entity.stock_management_entity.ItemTyrePresa;
 import org.rtss.mosad_backend.model_mapper.stock_management_model_mapper.ItemTyrePresaDTOMapper;
-import org.rtss.mosad_backend.repository.ItemTyrePresaRepo;
+import org.rtss.mosad_backend.repository.stock_management_repository.ItemTyrePresaRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
